Move ics file to separate mod
