+++ src/poll/finalize/emails.rs
@@ -1,5 +1,5 @@
 use crate::email::EmailMessage;
-use crate::event::{Event, EventEmailSender as DynEventEmailSender};
+use crate::event::{Event, EventEmailSender as DynEventEmailSender, Ics};
 use crate::fmt::LongEventTitle;
 use crate::uri::{HasUriBuilder as _, UriBuilder};
 use crate::users::User;
@@ -63,7 +63,7 @@ impl EventEmailSender {
         let event_url =
             uri!(auto_login(user, event.estimated_ends_at()); self.uri_builder, crate::event::event_page(id = event.id))
                 .await?;
-        let ics_file = crate::play::to_calendar(event, &self.uri_builder)?.to_string();
+        let ics_file = Ics::from_event(event, &self.uri_builder)?.0;
         let email = InvitedEmail {
             event,
             event_url,
